refactor(arch): decompose four oversized files into focused submodules
Split the four remaining non-test files flagged by #357 into cohesive
per-domain submodules. Pure structural refactor — every public path is
preserved (re-exported from each module root) and all behavior is moved
verbatim. Verified locally: cargo fmt, clippy -D warnings, and the full
534-test suite all pass.

- ndi/pipeline.rs (1204) -> pipeline/{build,lifecycle,consumers,tests}.rs
- ndi/manager.rs   (1099) -> manager/{lifecycle,supervisor,whep}.rs
- repository/bible.rs (1164) -> bible/{query,import,presentations,tests}.rs
- state/slides.rs  (1024) -> slides/{compose,edit_ops,tests}.rs

Note on the cap: the project's quality-check counts production lines only
(everything before the first #[cfg(test)]), so by the enforced metric all
four files were already under the 1000-line cap (861/874/788/556). The
arch-review issue counted raw wc -l (incl. exempt test modules). The split
still stands on maintainability grounds — these are 1000-1200 raw-line
files — and brings every production module comfortably under the target.

Closes #357
