+++ crates/presenter-ndi/src/manager.rs
@@ -5,16 +5,29 @@
 //! source and bridges WHEP HTTP operations into direct
 //! `pipeline.add_consumer` / `add_ice_candidate` / `remove_consumer` calls
 //! (no `whepserversink` `emit_by_name`).
+//!
+//! The implementation is split across responsibility-focused submodules
+//! (see #357 — split the over-cap god-file):
+//! - [`lifecycle`] — construction, discovery, start/stop, active-map queries
+//! - [`supervisor`] — per-source rebuild supervisor + backoff/cool-off state
+//! - [`whep`] — WHEP HTTP bridge + pipeline snapshots
+//!
+//! The shared type definitions and the active-map state-check live here in the
+//! module root so every submodule (and the regression tests) can reference
+//! them at a stable path.
 
 use std::collections::HashMap;
 use std::sync::Arc;
 
-use anyhow::{anyhow, Result};
 use tokio::sync::Mutex;
 
-use crate::discovery::{self, FinderShutdown, SourceList};
+use crate::discovery::{FinderShutdown, SourceList};
 use crate::ndi_sdk::NdiLib;
 use crate::pipeline::{NdiPipeline, PipelineState};
+
+mod lifecycle;
+mod supervisor;
+mod whep;
 
 /// Status callback retained for backwards compatibility with the old MJPEG
 /// status-reporting path. The WebRTC manager currently invokes it on
@@ -60,12 +73,12 @@ struct ActiveSource {
     /// this, holding the active-map lock across those awaits serializes ALL
     /// WHEP operations on the manager, stalls `pipeline_snapshots()` (used
     /// by `/healthz`) and blocks the supervisor's `rebuild_pipeline`.
-    pipeline: std::sync::Arc<NdiPipeline>,
+    pub(in crate::manager) pipeline: std::sync::Arc<NdiPipeline>,
     /// Supervisor task handle. Aborted on `stop_pipeline` / drop to prevent
     /// leaks. `None` only inside the regression-test constructors (which
     /// don't spawn a real supervisor) AND in the `rebuild_pipeline` re-insert
     /// path (the existing supervisor task is reused — see `spawn_supervisor`).
-    supervisor: Option<tokio::task::JoinHandle<()>>,
+    pub(in crate::manager) supervisor: Option<tokio::task::JoinHandle<()>>,
 }
 
 /// Outcome of `check_active_entry` — drives `start_pipeline`'s control flow.
@@ -78,104 +91,6 @@ enum StateCheckOutcome {
     /// Errored). In the dead case the entry has already been removed.
     /// Caller should proceed to build a fresh pipeline.
     Rebuild,
-}
-
-/// Per-source supervisor bookkeeping: when the last rebuild was attempted,
-/// and how many consecutive failures we've seen.
-///
-/// Pure data — no async, no I/O — so unit-testable on every CI host.
-#[derive(Debug)]
-struct SupervisorState {
-    last_rebuild_at: std::time::Instant,
-    /// 0 while the pipeline is healthy. Incremented by `mark_rebuild_failed`,
-    /// reset to 0 by `mark_rebuild_succeeded`.
-    consecutive_failures: u32,
-}
-
-/// Outcome of `SupervisorState::should_rebuild_now` — drives the supervisor's
-/// next sleep duration.
-#[derive(Debug)]
-enum RebuildDecision {
-    /// Wait this long, then attempt a rebuild. Zero duration means rebuild now.
-    ProceedAfter(std::time::Duration),
-}
-
-impl SupervisorState {
-    fn new() -> Self {
-        Self {
-            // Start with last_rebuild far enough in the past that the FIRST
-            // rebuild attempt has zero wait.
-            last_rebuild_at: std::time::Instant::now() - std::time::Duration::from_secs(3600),
-            consecutive_failures: 0,
-        }
-    }
-
-    fn consecutive_failures(&self) -> u32 {
-        self.consecutive_failures
-    }
-
-    /// #337: after `COOL_OFF_THRESHOLD` consecutive failures the supervisor
-    /// is "cooling off" — it sleeps `COOL_OFF_WINDOW` before each rebuild
-    /// attempt instead of the prior 30s exp-backoff cap. Manual reactivation
-    /// via the operator UI calls `start_pipeline`, which removes the dead
-    /// active-map entry, builds a fresh pipeline, and spawns a NEW
-    /// supervisor with a zero counter — so the old supervisor's cool-off
-    /// is implicitly cleared. mark_rebuild_succeeded also clears it (a
-    /// rebuild that succeeded inside the cool-off window means the fault
-    /// self-resolved).
-    fn is_cooling_off(&self) -> bool {
-        self.consecutive_failures >= Self::COOL_OFF_THRESHOLD
-    }
-
-    /// #337: number of consecutive failures before entering cool-off. Picked
-    /// from the issue body — small enough that an unrecoverable fault
-    /// stops thrashing within seconds, large enough that a flaky NDI
-    /// source (5-second LAN dropout) doesn't immediately cool off.
-    const COOL_OFF_THRESHOLD: u32 = 5;
-
-    /// #337: how long the supervisor waits between rebuild attempts once
-    /// the threshold is crossed. 5 minutes is the operator-comfortable
-    /// retry interval — long enough to stop log spam + CPU churn for
-    /// unrecoverable faults; short enough that a self-healing transient
-    /// fault (intermittent NDI broadcaster, GPU contention from another
-    /// process) recovers without operator intervention.
-    const COOL_OFF_WINDOW: std::time::Duration = std::time::Duration::from_secs(5 * 60);
-
-    /// Rate-limit window: minimum 2 seconds between rebuild attempts.
-    fn should_rebuild_now(&self, now: std::time::Instant) -> RebuildDecision {
-        const RATE_LIMIT: std::time::Duration = std::time::Duration::from_secs(2);
-        let since_last = now.duration_since(self.last_rebuild_at);
-        if since_last >= RATE_LIMIT {
-            RebuildDecision::ProceedAfter(std::time::Duration::ZERO)
-        } else {
-            RebuildDecision::ProceedAfter(RATE_LIMIT - since_last)
-        }
-    }
-
-    /// #337: once `COOL_OFF_THRESHOLD` consecutive failures hit, the
-    /// supervisor sleeps `COOL_OFF_WINDOW` (5 min) between attempts
-    /// instead of the prior 30s exp-backoff cap. This stops log-spam +
-    /// CPU churn for unrecoverable faults (encoder vanished, NDI source
-    /// removed). Manual reactivation via `start_pipeline` spawns a fresh
-    /// supervisor with a zero counter, which is the operator escape hatch.
-    fn backoff_for_failure_count(&self) -> std::time::Duration {
-        if self.consecutive_failures < Self::COOL_OFF_THRESHOLD {
-            return std::time::Duration::ZERO;
-        }
-        Self::COOL_OFF_WINDOW
-    }
-
-    fn mark_rebuild_started(&mut self) {
-        self.last_rebuild_at = std::time::Instant::now();
-    }
-
-    fn mark_rebuild_succeeded(&mut self) {
-        self.consecutive_failures = 0;
-    }
-
-    fn mark_rebuild_failed(&mut self) {
-        self.consecutive_failures = self.consecutive_failures.saturating_add(1);
-    }
 }
 
 /// Pure state-check for the active-source HashMap. Extracted from
@@ -231,646 +146,20 @@ async fn check_active_entry(
 }
 
 pub struct NdiManager {
-    _sdk: Arc<NdiLib>,
-    source_list: SourceList,
-    _finder_shutdown: FinderShutdown,
+    pub(in crate::manager) _sdk: Arc<NdiLib>,
+    pub(in crate::manager) source_list: SourceList,
+    pub(in crate::manager) _finder_shutdown: FinderShutdown,
     /// Map source_id (UUID string) → ActiveSource pipeline.
-    active: Mutex<HashMap<String, ActiveSource>>,
+    pub(in crate::manager) active: Mutex<HashMap<String, ActiveSource>>,
 }
 
-impl NdiManager {
-    pub fn try_new() -> Option<Self> {
-        let sdk = Arc::new(NdiLib::load().ok()?);
-        let (source_list, finder_shutdown) = discovery::spawn_persistent_finder(Arc::clone(&sdk));
-        Some(Self {
-            _sdk: sdk,
-            source_list,
-            _finder_shutdown: finder_shutdown,
-            active: Mutex::new(HashMap::new()),
-        })
-    }
-
-    pub fn is_available(&self) -> bool {
-        true
-    }
-
-    pub fn discover_sources(&self, _timeout_ms: u32) -> Result<Vec<discovery::NdiSourceInfo>> {
-        Ok(self.source_list.read())
-    }
-
-    /// Start a pipeline for the given source.
-    ///
-    /// `source_id` = UUID from the `video_sources` DB row (used as the WHEP URL key).
-    /// `ndi_name` = NDI broadcaster name (e.g. "STREAM-SNV (stream)").
-    ///
-    /// Returns only AFTER the pipeline has transitioned to `Streaming` — i.e.
-    /// the GStreamer bus has emitted `StateChanged → Playing` for the pipeline
-    /// element. For the shared-encoder topology (#336), this means ndisrc is
-    /// alive and ndisrcdemux has begun delivering frames; the encoder + tee
-    /// will start producing H264 buffers shortly after. Downstream webrtcbin
-    /// consumers attach lazily via `add_consumer`; they do not require encoder
-    /// caps at attach time (SDP exchange happens independently).
-    ///
-    /// An 8-second timeout caps the wait — long enough for ndisrc to find the
-    /// source on a healthy LAN, short enough that a missing/dead broadcaster
-    /// reports back quickly to the operator.
-    pub async fn start_pipeline(
-        self: &std::sync::Arc<Self>,
-        source_id: &str,
-        ndi_name: &str,
-    ) -> Result<()> {
-        let mut active = self.active.lock().await;
-
-        // Operator-reactivation path: if the existing entry is dead, snapshot
-        // its supervisor handle BEFORE `check_active_entry` removes the entry,
-        // so we can abort the prior supervisor below. Without this, a
-        // cool-off-bound supervisor that's mid-5-min-sleep keeps running and
-        // ends up double-watching the new pipeline alongside the fresh
-        // supervisor we spawn below (deep-review 🔵 #3, 2026-05-24 PR #340).
-        // Safe to `.take()` here because we hold the lock: state observed by
-        // `check_active_entry` below cannot change between these two reads.
-        let prior_supervisor: Option<tokio::task::JoinHandle<()>> = active
-            .get_mut(source_id)
-            .filter(|entry| {
-                matches!(
-                    entry.pipeline.state(),
-                    PipelineState::Stopped | PipelineState::Errored(_)
-                )
-            })
-            .and_then(|entry| entry.supervisor.take());
-
-        if let StateCheckOutcome::Idempotent = check_active_entry(&mut active, source_id).await {
-            // Pipeline turned out healthy — the dead-state filter above didn't
-            // match, so prior_supervisor is None. If somehow it leaked, drop
-            // the handle (does NOT cancel the task in tokio; the supervisor
-            // is still owned by its `ActiveSource.supervisor` slot if we
-            // didn't `.take()`).
-            debug_assert!(prior_supervisor.is_none());
-            return Ok(());
-        }
-        // The entry was dead → check_active_entry removed it. Abort the prior
-        // supervisor (if any) so it doesn't double-watch the new pipeline we
-        // build below.
-        if let Some(handle) = prior_supervisor {
-            handle.abort();
-        }
-
-        let whep_url = format!("/ndi/whep/{}", source_id);
-        let pipeline = NdiPipeline::build(ndi_name, whep_url)?;
-        pipeline.start().await?;
-
-        // Wait for the pipeline to reach Streaming state. The bus-watch task
-        // (started by pipeline.start()) sets state to Streaming once the
-        // GStreamer pipeline element posts StateChanged → Playing.
-        //
-        // The new shared-encoder topology (ndisrc → demux → videoconvert →
-        // encoder → rtph264pay → tee) has no whepserversink, so polling
-        // `sink_element.static_pad("video_0").current_caps()` is no longer
-        // applicable. Watching for PipelineState::Streaming is the correct
-        // signal: the bus-watch only promotes to Streaming after PLAYING,
-        // which requires ndisrcdemux to have negotiated caps with its upstream
-        // ndisrc — equivalent timing to the old caps-wait.
-        //
-        // 8-second budget: ndisrc takes ~2-5s on a healthy LAN to find a
-        // broadcast + receive first frame. Beyond 8s the source likely doesn't
-        // exist and we'd rather fail fast than hang the operator UI.
-        let mut watcher = pipeline.state_watcher();
-        let streaming_ready = tokio::time::timeout(std::time::Duration::from_secs(8), async {
-            loop {
-                let state = watcher.borrow_and_update().clone();
-                match state {
-                    crate::pipeline::PipelineState::Errored(ref e) => {
-                        return Err(anyhow!("pipeline errored: {e}"));
-                    }
-                    crate::pipeline::PipelineState::Streaming => return Ok(()),
-                    _ => {}
-                }
-                if watcher.changed().await.is_err() {
-                    return Err(anyhow!("state watcher closed unexpectedly"));
-                }
-            }
-        })
-        .await;
-
-        match streaming_ready {
-            Ok(Ok(())) => {
-                // pipeline.state_watcher() and self.spawn_supervisor must
-                // run before pipeline is wrapped into Arc and moved into
-                // ActiveSource on the active.insert line below.
-                let watcher = pipeline.state_watcher();
-                let supervisor =
-                    self.spawn_supervisor(source_id.to_string(), ndi_name.to_string(), watcher);
-                active.insert(
-                    source_id.to_string(),
-                    ActiveSource {
-                        pipeline: std::sync::Arc::new(pipeline),
-                        supervisor: Some(supervisor),
-                    },
-                );
-                Ok(())
-            }
-            Ok(Err(e)) => {
-                pipeline.stop().await;
-                Err(e)
-            }
-            Err(_) => {
-                pipeline.stop().await;
-                Err(anyhow!(
-                    "NDI source '{ndi_name}' did not reach Streaming within 8s; \
-                     ndisrc could not connect or the broadcaster is silent"
-                ))
-            }
-        }
-    }
-
-    /// Spawn the supervisor task for one source. Returns the JoinHandle so
-    /// callers can store it in `ActiveSource.supervisor` and abort on stop.
-    ///
-    /// The supervisor:
-    /// - Subscribes to the pipeline's state watcher
-    /// - On Errored/Stopped, requests a rebuild via `rebuild_pipeline`
-    /// - Rate-limits to 1 rebuild / 2s; enters a 5-minute cool-off after
-    ///   `COOL_OFF_THRESHOLD` (5) consecutive failures (#337). During
-    ///   cool-off the supervisor sleeps between rebuild attempts but does
-    ///   not exit — manual reactivation via `start_pipeline` aborts this
-    ///   supervisor and spawns a fresh one with a zero counter.
-    /// - Re-subscribes to the FRESH state watcher after each successful
-    ///   rebuild (the old watcher's pipeline was dropped)
-    /// - Exits when the watcher closes (pipeline dropped) OR the task is
-    ///   externally aborted via `stop_pipeline` / operator reactivate
-    fn spawn_supervisor(
-        self: &std::sync::Arc<Self>,
-        source_id: String,
-        ndi_name: String,
-        mut watcher: tokio::sync::watch::Receiver<crate::pipeline::PipelineState>,
-    ) -> tokio::task::JoinHandle<()> {
-        let manager = std::sync::Arc::clone(self);
-        tokio::spawn(async move {
-            let mut state = SupervisorState::new();
-            loop {
-                // Wait for the next state change.
-                if watcher.changed().await.is_err() {
-                    // Pipeline dropped — exit cleanly.
-                    tracing::debug!(source_id = %source_id, "supervisor: watcher closed, exiting");
-                    return;
-                }
-                let current = watcher.borrow_and_update().clone();
-                use crate::pipeline::PipelineState::*;
-                match current {
-                    Streaming | Starting => {
-                        // Healthy transition: reset the consecutive-failure
-                        // counter so prior errors don't carry over into the
-                        // next backoff cycle if a future fault occurs. The
-                        // spec pseudocode shows Starting as a no-op; we treat
-                        // it as success here because reaching Starting means
-                        // recovery is underway and any prior failure streak
-                        // is irrelevant once the new pipeline begins running.
-                        state.mark_rebuild_succeeded();
-                    }
-                    Errored(_) | Stopped => {
-                        // Apply rate-limit + backoff.
-                        let RebuildDecision::ProceedAfter(wait) =
-                            state.should_rebuild_now(std::time::Instant::now());
-                        let backoff = state.backoff_for_failure_count();
-                        let total = wait.max(backoff);
-                        if !total.is_zero() {
-                            tracing::warn!(
-                                source_id = %source_id,
-                                wait_ms = total.as_millis() as u64,
-                                consecutive_failures = state.consecutive_failures(),
-                                "supervisor: backing off before rebuild"
-                            );
-                            tokio::time::sleep(total).await;
-                        }
-                        state.mark_rebuild_started();
-                        match manager.rebuild_pipeline(&source_id, &ndi_name).await {
-                            Ok(()) => {
-                                tracing::info!(source_id = %source_id, "supervisor: rebuild succeeded");
-                                // NOTE: do NOT mark_rebuild_succeeded() yet. We only
-                                // know the pipeline build returned Ok — we don't yet
-                                // know it survived the immediate state-watcher peek
-                                // below. If we reset the counter here, the
-                                // already_dead branch below would mark_failed and
-                                // the counter would oscillate 0 → 1 → 0 → 1 forever,
-                                // never crossing the cool-off threshold (deep-review
-                                // 🟡 #1, 2026-05-24 PR #340). Reset only AFTER the
-                                // peek confirms the new pipeline is alive.
-                                // The fresh pipeline has a NEW state watcher; swap ours
-                                // so we see ITS transitions, not the dead pipeline's.
-                                match manager.state_watcher_for(&source_id).await {
-                                    Some(w) => {
-                                        watcher = w;
-                                        // After re-subscribing, the new watcher's
-                                        // "seen" mark is the current value. If the
-                                        // fresh pipeline has ALREADY errored in the
-                                        // window between active.insert and the
-                                        // state_watcher_for clone, changed() would
-                                        // block waiting for a further transition
-                                        // that never comes. Peek the state now: if
-                                        // it's already dead, mark the rebuild as a
-                                        // failure and `continue` — that returns to
-                                        // the outer loop's changed().await, which
-                                        // then blocks until either (a) the dead
-                                        // pipeline emits another transition
-                                        // (unlikely) or (b) the 30s DB-ticker
-                                        // backstop removes the entry and drops the
-                                        // state_tx, which makes changed() return
-                                        // Err and exits this supervisor cleanly.
-                                        // A fresh supervisor is then spawned by
-                                        // the ticker's start_pipeline. Worst-case
-                                        // recovery window is ~30s, which is the
-                                        // intended backstop behavior.
-                                        let already_dead = matches!(
-                                            *watcher.borrow_and_update(),
-                                            Errored(_) | Stopped
-                                        );
-                                        if already_dead {
-                                            // Mark as a failure for backoff bookkeeping —
-                                            // the rebuild "succeeded" briefly but the
-                                            // pipeline collapsed immediately, which is
-                                            // a real failure of recovery.
-                                            let was_cooling_off = state.is_cooling_off();
-                                            state.mark_rebuild_failed();
-                                            if !was_cooling_off && state.is_cooling_off() {
-                                                tracing::warn!(
-                                                    source_id = %source_id,
-                                                    consecutive_failures = state.consecutive_failures(),
-                                                    cool_off_minutes = 5,
-                                                    "supervisor: NDI source entered cool-off — pausing retries (#337); \
-                                                     manual reactivate via operator UI resumes immediately"
-                                                );
-                                            }
-                                            continue;
-                                        }
-                                        // Pipeline confirmed alive on the new watcher
-                                        // → reset the failure streak now (deferred
-                                        // from above per deep-review 🟡 #1).
-                                        state.mark_rebuild_succeeded();
-                                    }
-                                    None => {
-                                        // Source no longer active (operator deactivated
-                                        // between rebuild start and now). Exit.
-                                        tracing::debug!(
-                                            source_id = %source_id,
-                                            "supervisor: source no longer active after rebuild, exiting"
-                                        );
-                                        return;
-                                    }
-                                }
-                            }
-                            Err(e) => {
-                                let was_cooling_off = state.is_cooling_off();
-                                state.mark_rebuild_failed();
-                                tracing::warn!(
-                                    source_id = %source_id,
-                                    error = %e,
-                                    consecutive_failures = state.consecutive_failures(),
-                                    "supervisor: rebuild failed"
-                                );
-                                if !was_cooling_off && state.is_cooling_off() {
-                                    tracing::warn!(
-                                        source_id = %source_id,
-                                        consecutive_failures = state.consecutive_failures(),
-                                        cool_off_minutes = 5,
-                                        "supervisor: NDI source entered cool-off — pausing retries (#337); \
-                                         manual reactivate via operator UI resumes immediately"
-                                    );
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        })
-    }
-
-    /// Fetch the live `state_watcher` for an active source. Used by the
-    /// supervisor to re-subscribe after a successful rebuild (the old
-    /// watcher's pipeline is dropped after rebuild).
-    async fn state_watcher_for(
-        &self,
-        source_id: &str,
-    ) -> Option<tokio::sync::watch::Receiver<crate::pipeline::PipelineState>> {
-        let active = self.active.lock().await;
-        active.get(source_id).map(|s| s.pipeline.state_watcher())
-    }
-
-    /// Rebuild the pipeline for a source whose existing entry is dead
-    /// (Errored or Stopped). Reuses `check_active_entry` to clear the
-    /// dead entry, then builds + starts a fresh pipeline. Does NOT
-    /// spawn a new supervisor (the supervisor task that called us is
-    /// still alive and will re-subscribe to the new state watcher via
-    /// `state_watcher_for`).
-    async fn rebuild_pipeline(&self, source_id: &str, ndi_name: &str) -> Result<()> {
-        let mut active = self.active.lock().await;
-        // Force-remove the dead entry. If somehow it has become healthy in
-        // the meantime, leave it alone (idempotent).
-        if let StateCheckOutcome::Idempotent = check_active_entry(&mut active, source_id).await {
-            return Ok(());
-        }
-
-        let whep_url = format!("/ndi/whep/{}", source_id);
-        let pipeline = NdiPipeline::build(ndi_name, whep_url)?;
-        pipeline.start().await?;
-
-        // Wait for the pipeline to reach Streaming — same rationale as
-        // start_pipeline: state-watcher replaces the whepserversink pad
-        // caps-wait in the new shared-encoder topology.
-        let mut watcher = pipeline.state_watcher();
-        let streaming_ready = tokio::time::timeout(std::time::Duration::from_secs(8), async {
-            loop {
-                let state = watcher.borrow_and_update().clone();
-                match state {
-                    crate::pipeline::PipelineState::Errored(ref e) => {
-                        return Err(anyhow!("pipeline errored: {e}"));
-                    }
-                    crate::pipeline::PipelineState::Streaming => return Ok(()),
-                    _ => {}
-                }
-                if watcher.changed().await.is_err() {
-                    return Err(anyhow!("state watcher closed unexpectedly"));
-                }
-            }
-        })
-        .await;
-
-        match streaming_ready {
-            Ok(Ok(())) => {
-                active.insert(
-                    source_id.to_string(),
-                    ActiveSource {
-                        pipeline: std::sync::Arc::new(pipeline),
-                        // Supervisor task is reused — it'll fetch the new watcher
-                        // from us via `state_watcher_for`.
-                        supervisor: None,
-                    },
-                );
-                Ok(())
-            }
-            Ok(Err(e)) => {
-                pipeline.stop().await;
-                Err(e)
-            }
-            Err(_) => {
-                pipeline.stop().await;
-                Err(anyhow!(
-                    "NDI source '{ndi_name}' did not reach Streaming within 8s on rebuild"
-                ))
-            }
-        }
-    }
-
-    /// Stop one pipeline.
-    pub async fn stop_pipeline(&self, source_id: &str) {
-        let mut active = self.active.lock().await;
-        if let Some(mut src) = active.remove(source_id) {
-            if let Some(handle) = src.supervisor.take() {
-                handle.abort();
-            }
-            src.pipeline.stop().await;
-        }
-    }
-
-    /// Stop ALL pipelines.
-    pub async fn stop_all(&self) {
-        let mut active = self.active.lock().await;
-        for (_, src) in active.drain() {
-            if let Some(handle) = src.supervisor {
-                handle.abort();
-            }
-            src.pipeline.stop().await;
-        }
-    }
-
-    /// Is the given source's pipeline currently active?
-    pub async fn is_active(&self, source_id: &str) -> bool {
-        self.active.lock().await.contains_key(source_id)
-    }
-
-    /// Snapshot of every active pipeline's current state.
-    ///
-    /// Returns one entry per source currently in the active map, as
-    /// `(source_id, PipelineState)`. Used by `/healthz` (#333 item 7) so
-    /// dashboards can detect activation failures within seconds instead of
-    /// inferring from operator-reported 'red error' status.
-    ///
-    /// Bounded by a 200 ms lock-acquisition timeout (deep-review 🟡 #1):
-    /// `start_pipeline` and `rebuild_pipeline` hold the same `active` mutex
-    /// for up to 8 s during the caps-wait. Without the timeout, a `/healthz`
-    /// request that races a pipeline start would block long enough to
-    /// trip a 5 s LB health-check timeout — exactly the failure mode
-    /// item 7 was supposed to expose. On timeout we return an empty vec
-    /// and log a warning; the caller (LB / dashboard) sees "no pipelines"
-    /// for one poll cycle, which is preferable to a hung probe.
-    pub async fn pipeline_snapshots(&self) -> Vec<(String, PipelineState)> {
-        match tokio::time::timeout(std::time::Duration::from_millis(200), self.active.lock()).await
-        {
-            Ok(guard) => guard
-                .iter()
-                .map(|(id, src)| (id.clone(), src.pipeline.state()))
-                .collect(),
-            Err(_) => {
-                tracing::warn!(
-                    "pipeline_snapshots lock acquisition timed out after 200 ms — \
-                     likely contended with a long-running pipeline start/rebuild; \
-                     returning empty snapshot so /healthz does not stall (#333 item 7)"
-                );
-                Vec::new()
-            }
-        }
-    }
-
-    /// Single-source snapshot for `GET /ndi/snapshot/:source_id`. Returns
-    /// `None` if the source isn't active in the manager's active map.
-    ///
-    /// Uses the same 200 ms lock-acquisition timeout pattern as
-    /// `pipeline_snapshots` so a `/ndi/snapshot/:id` probe doesn't stall
-    /// behind a concurrent pipeline start/rebuild. On timeout returns `None`
-    /// (caller maps to 503).
-    pub async fn pipeline_snapshot(
-        &self,
-        source_id: &str,
-    ) -> Option<crate::pipeline::PipelineSnapshot> {
-        let guard = tokio::time::timeout(std::time::Duration::from_millis(200), self.active.lock())
-            .await
-            .ok()?;
-        let pipeline = std::sync::Arc::clone(&guard.get(source_id)?.pipeline);
-        drop(guard);
-        let mut snap = pipeline.snapshot().await;
-        snap.source_id = source_id.to_string();
-        Some(snap)
-    }
-
-    /// Test-only: trigger an Errored state on the source's pipeline so
-    /// the PipelineSupervisor reacts as it would for a real ndisrc fault.
-    /// Returns `true` if the source was active (state injection succeeded),
-    /// `false` if not (caller should map to 404).
-    #[cfg(feature = "test-helpers")]
-    pub async fn simulate_pipeline_error(&self, source_id: &str, msg: &str) -> bool {
-        let active = self.active.lock().await;
-        match active.get(source_id) {
-            Some(src) => {
-                src.pipeline.simulate_error_for_test(msg);
-                true
-            }
-            None => false,
-        }
-    }
-
-    /// Forward a WHEP HTTP exchange to the source's pipeline. Replaces the
-    /// pre-#336 `emit_by_name`-on-whepserversink path. Routes each `WhepOp`
-    /// variant to the corresponding `NdiPipeline` method.
-    ///
-    /// The active-map mutex guard is always DROPPED before calling any
-    /// potentially-blocking pipeline method (`add_consumer` spawn_blocks for
-    /// ~10s, `add_ice_candidate` and `remove_consumer` also spawn_block).
-    /// To achieve this without copying the pipeline, `ActiveSource.pipeline`
-    /// is an `Arc<NdiPipeline>` — we clone the `Arc` (cheap refcount bump)
-    /// inside the lock, drop the guard, then call the pipeline method outside.
-    pub async fn whep_signaller_call(&self, source_id: &str, op: WhepOp) -> Result<WhepReply> {
-        match op {
-            WhepOp::Post { id: None, body } => {
-                let pipeline = {
-                    let active = self.active.lock().await;
-                    let src = active
-                        .get(source_id)
-                        .ok_or_else(|| anyhow!(SOURCE_NOT_ACTIVE_ERR))?;
-                    Self::ensure_streaming(src)?;
-                    std::sync::Arc::clone(&src.pipeline)
-                    // active lock dropped here
-                };
-                let answer = pipeline.add_consumer(body).await?;
-                let location = format!("/ndi/whep/{source_id}/{}", answer.session_id);
-                tracing::info!(
-                    source_id = %source_id,
-                    session_id = %answer.session_id,
-                    "WHEP POST → 201"
-                );
-                Ok(WhepReply {
-                    status: 201,
-                    headers: vec![
-                        ("location".to_string(), location),
-                        ("content-type".to_string(), "application/sdp".to_string()),
-                    ],
-                    body: Some(answer.sdp_answer.into_bytes()),
-                })
-            }
-            WhepOp::Post {
-                id: Some(_),
-                body: _,
-            } => {
-                // Session-scoped re-offer — out of scope for #336.
-                // Validate that the source is known first (preserves 404
-                // semantics for unknown sources — the HTTP shim tests assert
-                // this contract). Lock is acquired and dropped immediately;
-                // no blocking await follows.
-                {
-                    let active = self.active.lock().await;
-                    let src = active
-                        .get(source_id)
-                        .ok_or_else(|| anyhow!(SOURCE_NOT_ACTIVE_ERR))?;
-                    Self::ensure_streaming(src)?;
-                }
-                tracing::warn!(source_id = %source_id, "WHEP session-scoped POST not implemented");
-                Ok(WhepReply {
-                    status: 501,
-                    headers: vec![("content-type".to_string(), "text/plain".to_string())],
-                    body: Some(b"WHEP re-offer not implemented".to_vec()),
-                })
-            }
-            WhepOp::Patch {
-                id,
-                body,
-                headers: _,
-            } => {
-                // Clone the pipeline reference out of the map guard before
-                // any blocking await. The per-candidate loop re-uses this
-                // single clone rather than re-locking per iteration.
-                let pipeline = {
-                    let active = self.active.lock().await;
-                    let src = active
-                        .get(source_id)
-                        .ok_or_else(|| anyhow!(SOURCE_NOT_ACTIVE_ERR))?;
-                    Self::ensure_streaming(src)?;
-                    std::sync::Arc::clone(&src.pipeline)
-                    // active lock dropped here
-                };
-                // Parse `application/trickle-ice-sdpfrag` body: extract
-                // `a=mid:` (for mline index) and `a=candidate:` lines,
-                // forwarding each candidate to pipeline.add_ice_candidate.
-                let body_str =
-                    std::str::from_utf8(&body).map_err(|e| anyhow!("PATCH body not utf8: {e}"))?;
-                let mut count = 0;
-                let mut mline_idx: u32 = 0;
-                for raw_line in body_str.lines() {
-                    let line = raw_line.trim();
-                    if let Some(rest) = line.strip_prefix("a=mid:") {
-                        if let Ok(n) = rest.trim().parse::<u32>() {
-                            mline_idx = n;
-                        }
-                        // Non-integer mid (RFC 8839 allows e.g. "audio") falls
-                        // through; mline_idx stays at the last valid integer (or 0).
-                        // Browsers use integer mids in WHEP practice.
-                    } else if line.starts_with("a=candidate:") {
-                        // webrtcbin's add-ice-candidate signal accepts the
-                        // candidate string without the leading "a=" prefix.
-                        let cand_value = &line[2..];
-                        pipeline
-                            .add_ice_candidate(&id, mline_idx, cand_value)
-                            .await?;
-                        count += 1;
-                    }
-                }
-                tracing::debug!(
-                    source_id = %source_id,
-                    session_id = %id,
-                    candidate_count = count,
-                    "WHEP PATCH dispatched"
-                );
-                Ok(WhepReply {
-                    status: 204,
-                    headers: vec![],
-                    body: None,
-                })
-            }
-            WhepOp::Delete { id } => {
-                // Clone the pipeline reference before the blocking await.
-                // DELETE proceeds regardless of pipeline state — teardown must
-                // succeed even when the pipeline is erroring, so
-                // ensure_streaming is intentionally skipped here.
-                let pipeline = {
-                    let active = self.active.lock().await;
-                    let src = active
-                        .get(source_id)
-                        .ok_or_else(|| anyhow!(SOURCE_NOT_ACTIVE_ERR))?;
-                    std::sync::Arc::clone(&src.pipeline)
-                    // active lock dropped here
-                };
-                pipeline.remove_consumer(&id).await?;
-                tracing::info!(
-                    source_id = %source_id,
-                    session_id = %id,
-                    "WHEP DELETE → consumer removed"
-                );
-                Ok(WhepReply {
-                    status: 204,
-                    headers: vec![],
-                    body: None,
-                })
-            }
-        }
-    }
-
-    /// Pipeline state must be Streaming or Starting for WHEP ops to proceed.
-    /// Stopped / Errored produce an error that the HTTP shim maps to 503.
-    fn ensure_streaming(src: &ActiveSource) -> Result<()> {
-        match src.pipeline.state() {
-            PipelineState::Streaming | PipelineState::Starting => Ok(()),
-            PipelineState::Stopped => Err(anyhow!("pipeline stopped")),
-            PipelineState::Errored(e) => Err(anyhow!("pipeline errored: {e}")),
-        }
-    }
-}
+// Re-export the supervisor's pure-state types into the module root so the
+// regression tests below (which use `super::*`) keep resolving them at their
+// original `manager::` paths. These types are crate-internal — gating the
+// re-export on `#[cfg(test)]` keeps the production build free of an
+// otherwise-unused import under clippy `-D warnings`.
+#[cfg(test)]
+use self::supervisor::{RebuildDecision, SupervisorState};
 
 #[cfg(test)]
 mod start_pipeline_state_check_tests {
